Project Page created
All of the projects are pulled from the databse and presented in a table with view and edit buttons corresponding to the correct project IDs in the databse. In the table presenting all the projects the description has been set to 50 characters adding '...' to the end to show there is more description about the project if the click view.