Slugs Complete
All of the slugs have been linked with the routes and url generators. This allows for the project idea to be pulled in from the database and a url will be created based on the 'project-bazaar/' and the slug inputted hen creating the project. There was an error that was fixed which was the unique validation of the slug. it wouldnt allow updating of the project idea because the slug wasnt unique. This was fixed by creating an if statement to not validate the slug if it is being updated but do validate when a new project idea is being created.