Authentication
Authentication has been set up allowing users that are logged in to access the proijects pages to CRUD. If the user isnt logged in they cannnot access the page it redirects them to a log in page. Also the right hand dropdown nav is hidden unless logged in. The reset password and emails have been set up allowing a reset email link to be sent via email and works successfully all links up to the database.