Comments View
Now tha the comments are submitted to the databse the view have been created pulling the comments in using a @foreach. had an issue will sytax error which was an incorrect typo on my behalf trying to get to App\Commetns instead of App\Comment in the Project.php file. this threw and Exeption Error not loading the page. Would be good to implement 404 500 Error pages for the users