Comments Back-End
Comments can now be editted and deleted. A 'disabled' attribute had been set to the name and email fields in the edit.blade to make sure the author of the comment isnt changed. An auth middleware has also been set to the CommentsController to make sure only users logged in can edit or delete comments. A delete blade has been created to create and interface for the user deleting the comment.