Select Dropdown
The message "select course' was being repeated after every course option available. The issue was that the <option> tag was inside the @foreach loop. This issue was resolved by putting this before the @foreach loop meaning it is shown first and not looped everytime there is a new course shown.